Avoid ferdium to autostart after boot

+++ monitor60mhz.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 kscreen-doctor output.1.mode.2
-#sleep 0.2
-flatpak kill org.ferdium.Ferdium || true
-flatpak run org.ferdium.Ferdium --hidden &
+
+if pgrep -f "org.ferdium.Ferdium" > /dev/null; then
+    flatpak kill org.ferdium.Ferdium || true
+    flatpak run org.ferdium.Ferdium --hidden &
+fi
